requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
cffi==1.10.0
click==6.7
decorator==4.0.11
Django==1.11.1
django-rest-auth==0.9.1
djangorestframework==3.6.3
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
mysqlclient==1.3.10
numpy==1.12.1
packaging==16.8
passlib==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pycparser==2.17
Pygments==2.2.0
PyMySQL==0.7.11
pyparsing==2.2.0
python-dotenv==0.6.4
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
